ai-service/requirements.txt
flask==2.3.3
transformers==4.33.2
torch==2.0.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
